[]: Moving recovery code APIs under /identity and using POST
