+++ crates/standalone/src/routes/mod.rs
@@ -16,7 +16,10 @@ where
             "/database",
             database::control_routes(ctx.clone()).merge(database::worker_routes(ctx.clone())),
         )
-        .nest("/identity", identity::router(ctx.clone()))
+        .nest(
+            "/identity",
+            identity::control_routes(ctx.clone()).merge(identity::router(ctx.clone())),
+        )
         .nest("/energy", energy::router())
         .nest("/prometheus", prometheus::router())
         .nest("/metrics", metrics::router());
